Changed (int)(something + 0.5)/1000 to (float)(int)(something + 0.5)/1000. This ensures that the division is handled as a float and not as an int.

+++ initialize_atmos.c
@@ -73,7 +73,7 @@ void initialize_atmos(atmos_data_struct        *atmos,
   07-May-04 Replaced
 		rint(something)
 	    with
-		(int)(something + 0.5)
+		(float)(int)(something + 0.5)
 	    to handle rounding errors without resorting to rint()
 	    function.						TJB
 
@@ -416,7 +416,7 @@ void initialize_atmos(atmos_data_struct        *atmos,
 	  atmos[rec].vp[i] = forcing_data[VP][idx];
 
 
-	  atmos[rec].vp[i] = ((int)(atmos[rec].vp[i]*1000 + 0.5)/1000);
+	  atmos[rec].vp[i] = ((float)(int)(atmos[rec].vp[i]*1000 + 0.5)/1000);
 
 
 	  atmos[rec].vpd[i] = (svp(atmos[rec].air_temp[i])
